feat: move lang outside of index
